Add section "Parte de la Comunidad"
This commit adds the section "Parte de la comunidad". This section
lists the groups to which it belongs datamx.
Also it add the opendatamx logo in the section previously mentioned.

We can refactor the follow sections for take advantage of Jinja
templates system:

  - Nuestros Aliados
  - Con apoyo de
  - Parte de la comunidad

So as avoid create many tags and take the list model:

  [organization, name, logo_path]

Signed-off-by: Miguel Angel Gordian <miguel.angel@codeandomexico.org>